refseq transcript loading
